backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
Pillow>=10.0.0
pydantic>=2.0.0
